New hold/release mode for initializing the wave function
Update the GLSL FFT implementation with a more optimized version.
Fix a long-standing issue where the cursor gets stuck in "draw mode" when
holding down the input device and moving the cursor outside the canvas.
Rename and rearrange some of the GUI controls to hopefully make their
function more apparent.
In the instructions add documentation to some of the GUI controls which were
lacking in previous versions. Proofread the README file.
Add the 3840x2160 simulation grid size.