Fix some issues with mouse input
Remove behaviour that caused the canvas to be dragged in Firefox
whenever one holds down the mouse to place a new wave function or
potential barrier. Change how the initial momentum is computed
when using hold/release mode. Also fix an issue that caused unwanted
additional sketches to be created when drawing a potential barrier
for a non-square canvas. Remove commented-out markdown in the README.